pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "django-input-collection"
dynamic = ["version"]
description = "Input system driveable by any interaction process, such as a request-response cycle, API, or a Python function."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Pivotal Energy Solutions", email = "steve@pivotal.energy" },
]
keywords = [
    "django",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "django>=5.0",
    "djangorestframework",
]

[project.optional-dependencies]
test = [
    "factory-boy",
    "django-environ",
    "mysqlclient",
    "coverage",
    "pre-commit",
    "black",
    "bandit",
    "ruff"
]

[project.urls]
Homepage = "https://github.com/pivotal-energy-solutions/django-input-collection"
Issues = "https://github.com/pivotal-energy-solutions/django-input-collection/issues"
Download = "https://github.com/pivotal-energy-solutions/django-input-collection"
Thanks = "https://saythanks.io/to/rh0dium"
Source = "https://github.com/pivotal-energy-solutions/django-input-collection/"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = [
    "/django_input_collection",
    "/django_input_collection/templates/**/*",
]

[tool.hatch.build.targets.wheel]
packages = ['django_input_collection']
include = [
    "/django_input_collection/templates/**/*",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '(\.git|.venv|_build|build|dist|.*\/__pycache__\/)'

[tool.ruff]
line-length = 100
lint.ignore = ["F401"]

[tool.bandit]
targets = ['django_input_collection']
exclude_dirs = ["django_input_collection/tests", "demo_app"]
skips = ["B101", "B110", "B307", "B308", "B703"]

[tool.coverage.run]
branch = true
command_line = "demo_app/manage.py test --noinput --settings=demo_app.settings_test django_input_collection"
omit = [
    "*/demo_app/**",
    "*/migrations/*",
    "*/tests/**",
]

[tool.coverage.report]
fail_under = 66
precision = 1
skip_covered = true
skip_empty = true
ignore_errors = true
sort = "cover"

[tool.bumper]
exclude = [".idea", ".github", "demo_app"]
version_files = ["django_input_collection/__init__.py"]
repo = "pivotal-energy-solutions/django-input-collection"
report = "out.json"
